refactor(client): removing git folder in hooks
