+++ packages/amplication-client/src/Application/git/hooks/useGitCreate.tsx
@@ -4,9 +4,9 @@ import {
   EnumSourceControlService,
   GitRepo,
   RepoCreateInput,
-} from "../../../../models";
-import { useTracking } from "../../../../util/analytics";
-import { formatError } from "../../../../util/error";
+} from "../../../models";
+import { useTracking } from "../../../util/analytics";
+import { formatError } from "../../../util/error";
 
 type Props = {
   appId: string;
